Item small fixes
- fixed stale override names on Spear.prefab
- updated ItemController.Detach() to disable any active overlay to avoid item overlays getting stuck
- fixed Arm.prefab not allowing detection of collision w/ kinematic/static objects
- removed unnecessary rigid body from DoorSecret.prefab and decreased health
	- converted Health.maxHP to a float
- fixed DoorController typo from previous commit that could cause too-short combination strings

+++ Assets/Scripts/Health.cs
@@ -10,7 +10,7 @@ public class Health : MonoBehaviour
 	/// <summary>
 	/// The maximum hit points for the entity.
 	/// </summary>
-	public int maxHP = 1;
+	public float maxHP = 1.0f;
 
 	public GameObject m_healthUIParent;
 	public Sprite m_healthSprite;
@@ -121,7 +121,7 @@ public class Health : MonoBehaviour
 	private bool IncrementInternal(float diff)
 	{
 		float hpPrev = currentHP;
-		currentHP = Mathf.Clamp(currentHP + diff, 0, maxHP);
+		currentHP = Mathf.Clamp(currentHP + diff, 0.0f, maxHP);
 		return !Mathf.Approximately(currentHP, hpPrev);
 	}
